requirements.txt
greenlet==3.2.1
playwright==1.52.0
pyee==13.0.0
python-dotenv==1.1.0
typing_extensions==4.13.2
